Added comments and optimizations
Following up on the commits I’ve helped Adam push recently, I went over
our code to make sure everything was thoroughly commented, correctly
formatted, and optimally efficient.

A notable change, I’ve rewritten the checkCart() function to now return
the object that the function was looking for, rather than just to check
if it exists, forcing the dev to write another loop to actually find
and do something to the item in cart.